Initial commit
This is the initial commit for the reworked version of Herelink QGC. The rework was done such that fewer merge conflicts would happen when pulling new upstream changes.

The cmake build is still not fully set up for custom android builds. The followed files are modified in the upstream source:
* android/AndroidManifest.xml
* android/build.gradle
* CMakeLists.txt